Centralize loading / error tests into useDataFromServer tests
