+++ client/src/Pages/Courses/Tests/Courses-display.test.tsx
@@ -1,0 +1,22 @@
+/* eslint-disable max-lines-per-function */
+import { fireEvent } from '@testing-library/react';
+import App from 'Pages/App/App';
+import React from 'react';
+import { renderWithRouterAndProvider } from 'TestUtils/renderWith';
+
+test('Renders four courses for non-error server response', async () => {
+  // Note: mocked server response is handled by msw, in the src/mocks folder
+  // and src/setupTests.js. The handler is set to return
+  // TestUtils/Data/testCoursesData (see above) for /api/courses
+
+  // render entire App so that we can check Loading and Error
+  const screen = renderWithRouterAndProvider(<App />);
+
+  // click the 'courses' tab to trigger the Courses retrieval
+  const coursesNavLink = screen.getByRole('tab', { name: /courses/i });
+  fireEvent.click(coursesNavLink);
+
+  // check titles (all fake names start with "Course")
+  const titles = await screen.findAllByText(/Course \d/);
+  expect(titles.length).toBe(2);
+});
